refactor : LottoGenerator 관련 패키지 이동 (domain.generator)
