+++ src/main/java/lotto/model/generator/LottoGenerator.java
@@ -1,0 +1,8 @@
+package lotto.model.generator;
+
+import lotto.model.LottoPaper;
+import lotto.model.Lottos;
+
+public interface LottoGenerator {
+    Lottos generateLottos(LottoPaper lottoPaper);
+}
